dividing go_ib into multiple subfiles, each with one data structure
