+++ .vscode/settings.json
@@ -62,6 +62,7 @@
 		"text_encoding": "cpp",
 		"cinttypes": "cpp",
 		"typeinfo": "cpp",
-		"variant": "cpp"
+		"variant": "cpp",
+		"*.m": "cpp"
 	}
 }
